requirements.txt
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
channels==0.17.2
daphne==0.15.0
Django==1.10.1
django-cors-middleware==1.3.1
django-filter==0.14.0
djangorestframework==3.4.6
google-api-python-client==1.5.5
httplib2==0.9.2
jsonpickle==0.9.3
Markdown==2.6.6
msgpack-python==0.4.8
oauth2client==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
redis==2.10.5
rsa==3.4.2
six==1.10.0
Twisted==16.4.0
txaio==2.5.1
uritemplate==3.0.0
zope.interface==4.3.2
